Refactor common test inner classes into top level
