+++ subprojects/jdeferred-core/src/test/java/org/jdeferred/impl/SettleTest.java
@@ -24,7 +24,6 @@ import org.jdeferred.multiple.OneReject;
 import org.jdeferred.multiple.OneResult;
 import org.junit.Test;
 
-import java.security.SecureRandom;
 import java.util.concurrent.Callable;
 
 import static org.junit.Assert.assertEquals;
@@ -80,47 +79,6 @@ public class SettleTest extends AbstractDeferredTest {
 				assertTrue("OneOf at index " + i + " should be of type OneReject", values[0].get(i) instanceof OneReject);
 				assertTrue("Value at index " + i + " should be of type IndexedRuntimeException", values[0].get(i).getValue() instanceof IndexedRuntimeException);
 			}
-		}
-	}
-
-	private static class ResolvingCallable implements Callable<Integer> {
-		protected static final SecureRandom RANDOM = new SecureRandom();
-		protected final int index;
-
-		public ResolvingCallable(int index) {
-			this.index = index;
-		}
-
-		@Override
-		public Integer call() throws Exception {
-			Thread.sleep(500);
-			Thread.sleep(RANDOM.nextInt(100));
-			return index;
-		}
-	}
-
-	private static class RejectingCallable extends ResolvingCallable {
-		public RejectingCallable(int index) {
-			super(index);
-		}
-
-		@Override
-		public Integer call() throws Exception {
-			Thread.sleep(500);
-			Thread.sleep(RANDOM.nextInt(100));
-			throw new IndexedRuntimeException(index);
-		}
-	}
-
-	private static class IndexedRuntimeException extends RuntimeException {
-		private final int index;
-
-		private IndexedRuntimeException(int index) {
-			this.index = index;
-		}
-
-		public int getIndex() {
-			return index;
 		}
 	}
 }
